added key hashing
