+++ src/js/ui/NewDataExploration/WorkspaceLayers.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Layer from "./Layer";
 import {componentIsRendering} from "../TabSystem";
-import {randomizeIndex} from "./Helpers";
+import {hashIndex} from "./Helpers";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -15,7 +15,7 @@ function createWorkspace(layers, graphableLayers, openLayers, setOpenLayers, wor
     workspace.forEach((layer, index) => {
         if(layer) {
             const originalIndex = index;
-            index = randomizeIndex();
+            index = hashIndex(7) + index;
             workspaceLayers.push(
                 <div key={index} id={`layer-div-${originalIndex}`}>
                     <Layer layer={layers[originalIndex]} layerIndex={originalIndex} graphableLayers={graphableLayers} layerTitles={layerTitles}
@@ -38,7 +38,7 @@ export default function WorkspaceLayers(props) {
     return (
         <div className={classes.root}>
             {workspaceLayers.map((layer, index) => {
-                    index = randomizeIndex();
+                    index = hashIndex(11) + index;
                     return (<div key={index}>{layer}</div>)
                 }
             )}
